rename joint key retrieved to joint key upload
